First pass.
