+++ GitHubBugReport.Core/Issues/Models/DataModelIssue.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 
-namespace BugReport.DataModel
+namespace GitHubBugReport.Core.Issues.Models
 {
     public class DataModelIssue
     {
@@ -15,15 +14,12 @@ namespace BugReport.DataModel
         public Label[] Labels;
         public User User;
         public string HtmlUrl;
-
         public DateTimeOffset? CreatedAt;
         public DateTimeOffset? UpdatedAt;
         public DateTimeOffset? ClosedAt;
         public User ClosedBy;
-
         public PullRequest PullRequest;
         public Milestone Milestone;
-
 
         private Repository _repo;
         public Repository Repo
@@ -34,18 +30,15 @@ namespace BugReport.DataModel
                 {
                     _repo = Repository.FromHtmlUrl(HtmlUrl);
                 }
+
                 return _repo;
             }
         }
 
-        public bool IsIssueOrComment
-        {
-            get => (PullRequest == null);
-        }
-        public bool IsPullRequest
-        {
-            get => (PullRequest != null);
-        }
+        public bool IsIssueOrComment => (PullRequest == null);
+
+        public bool IsPullRequest => (PullRequest != null);
+
         public bool IsIssueKind(IssueKindFlags flags)
         {
             if (IsIssueOrComment)
@@ -59,18 +52,14 @@ namespace BugReport.DataModel
                     return ((flags & IssueKindFlags.Issue) != 0);
                 }
             }
+
             Debug.Assert(IsPullRequest);
             return ((flags & IssueKindFlags.PullRequest) != 0);
         }
 
-        public bool IsOpen
-        {
-            get => (State == Octokit.ItemState.Open);
-        }
-        public bool IsClosed
-        {
-            get => (State == Octokit.ItemState.Closed);
-        }
+        public bool IsOpen => (State == Octokit.ItemState.Open);
+
+        public bool IsClosed => (State == Octokit.ItemState.Closed);
 
         public bool HasLabel(string labelName) => Labels.Contains_ByName(labelName);
 
@@ -98,13 +87,13 @@ namespace BugReport.DataModel
             sw.WriteLine("Number: {0}", Number);
             sw.WriteLine("Type: {0}", (PullRequest == null) ? "Issue" : "PullRequest");
             sw.WriteLine("URL: {0}", HtmlUrl);
-            sw.WriteLine("State: {0}", State);
+            sw.WriteLine((string) "State: {0}", (object) State);
             sw.WriteLine("Assignee.Name:  {0}", (Assignee == null) ? "<null>" : Assignee.Name);
             sw.WriteLine("        .Login: {0}", (Assignee == null) ? "<null>" : Assignee.Login);
             sw.WriteLine("Labels.Name:");
             foreach (Label label in Labels)
             {
-                sw.WriteLine("    {0}", label.Name);
+                sw.WriteLine((string) "    {0}", (object) label.Name);
             }
             sw.WriteLine("Title: {0}", Title);
             //sw.WriteLine("Milestone.Title: {0}", (issue.Milestone == null) ? "<null>" : issue.Milestone.Title);
@@ -127,72 +116,5 @@ namespace BugReport.DataModel
             // Check also HtmlUrl which encodes repo
             return ((Number == issue.Number) && (HtmlUrl == issue.HtmlUrl));
         }
-    }
-
-    [FlagsAttribute]
-    public enum IssueKindFlags
-    {
-        Issue = 1,
-        PullRequest = 2,
-        Comment = 4,
-        All = Issue | PullRequest | Comment
-    }
-
-    public class Label
-    {
-        public string Name;
-
-        public Label(string name)
-        {
-            Name = name;
-        }
-
-        public bool Equals(string name)
-        {
-            return NameEqualityComparer.Equals(Name, name);
-        }
-
-        public static StringComparer NameEqualityComparer = StringComparer.InvariantCultureIgnoreCase;
-    }
-
-    public class User
-    {
-        public string Name;
-        public string Login;
-        public string HtmlUrl;
-    }
-
-    public class Milestone
-    {
-        public int Number;
-        public string Title;
-        public string Description;
-        public int OpenIssues;
-        public int ClosedIssues;
-        public Octokit.ItemState State;
-        public User Creator;
-        public DateTimeOffset CreatedAt;
-        public DateTimeOffset? DueOn;
-        public DateTimeOffset? ClosedAt;
-
-        public bool Equals(string title)
-        {
-            return TitleComparer.Equals(Title, title);
-        }
-
-        public static StringComparer TitleComparer = StringComparer.InvariantCultureIgnoreCase;
-    }
-
-    public class PullRequest
-    {
-        //public string HtmlUrl;
-
-        /*
-        public DateTimeOffset? CreatedAt;
-        public DateTimeOffset? UpdatedAt;
-        public DateTimeOffset? ClosedAt;
-        public DateTimeOffset? MergedAt;
-        public bool Merged;
-        */
     }
 }
