+++ GitHubBugReport.Core/Issues/Models/PullRequest.cs
@@ -1,0 +1,15 @@
+﻿namespace GitHubBugReport.Core.Issues.Models
+{
+    public class PullRequest
+    {
+        //public string HtmlUrl;
+
+        /*
+        public DateTimeOffset? CreatedAt;
+        public DateTimeOffset? UpdatedAt;
+        public DateTimeOffset? ClosedAt;
+        public DateTimeOffset? MergedAt;
+        public bool Merged;
+        */
+    }
+}
