+++ GitHubBugReport.Core/Issues/Models/User.cs
@@ -1,0 +1,9 @@
+namespace GitHubBugReport.Core.Issues.Models
+{
+    public class User
+    {
+        public string Name;
+        public string Login;
+        public string HtmlUrl;
+    }
+}
